Project 2 Extending the JWKS server/requirements.txt
cryptography==41.0.4
pyjwt==2.8.0
